Use int64 for the provisioning stack ID
provisioning.NewClient now takes the stack ID as int64 and treats it
as required (checkCloudLogin already guarantees it). The int32
X-Stack-Id boundary is confined to v6.SetStackID, which validates the
range; the provisioning endpoints carry it as a decimal-string header.

Removes the toInt32 stack-id conversions that leaked into
cmd/outputs_cloud.go and output/cloud/output.go, so those layers work
purely in int64. Resource IDs were already int64.

+++ internal/cloudapi/provisioning/client_test.go
@@ -16,7 +16,7 @@ func TestNewClient(t *testing.T) {
 	token := "test-token"
 	host := "https://api.k6.io"
 	version := "0.1.0"
-	stackID := int32(123)
+	stackID := int64(123)
 	timeout := 30 * time.Second
 
 	c, err := NewClient(logger, token, host, version, stackID, timeout)
